# Conflicts:
#	nts.uk/uk.com/com.ctx/sys/portal/nts.uk.ctx.sys.portal.ac/.classpath

+++ nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.infra/src/main/java/nts/uk/ctx/at/record/infra/entity/workrecord/closure/KclmtClosureHist_.java
@@ -20,10 +20,10 @@ public class KclmtClosureHist_ {
 	public static volatile SingularAttribute<KclmtClosureHist, String> name;
 	
 	/** The str D. */
-	public static volatile SingularAttribute<KclmtClosureHist, Long> strD;
+	public static volatile SingularAttribute<KclmtClosureHist, Integer> strD;
 	
 	/** The end D. */
-	public static volatile SingularAttribute<KclmtClosureHist, Long> endD;
+	public static volatile SingularAttribute<KclmtClosureHist, Integer> endD;
 	
 	/** The close day. */
 	public static volatile SingularAttribute<KclmtClosureHist, Integer> closeDay;
